Fix: EIntOverflow in TestMixedSizeRandomPattern in AdvancedTest.dpr
Resolved an EIntOverflow in the TestMixedSizeRandomPattern procedure in AdvancedTest.dpr. The overflow occurred during the Seed calculation due to intermediate multiplication exceeding the capacity of a 32-bit signed integer in debug builds.

The fix involves casting 'Seed' to 'Int64' before multiplication to force 64-bit arithmetic, preventing the overflow.

+++ Tests/Advanced/AdvancedTest.dpr
@@ -1446,7 +1446,7 @@ begin
   // Simple PRNG for reproducibility
   for i := 0 to Count - 1 do
   begin
-    Seed := (Seed * 1103515245 + 12345) and $7FFFFFFF;
+    Seed := (Int64(Seed) * 1103515245 + 12345) and $7FFFFFFF;
     Sizes[i] := (Seed mod 100000) + 1; // 1 to 100000 bytes
     Ptrs[i] := nil;
   end;
